review: separate errors
