+++ bin/faucet/src/pow/challenge.rs
@@ -2,10 +2,10 @@ use std::time::Duration;
 
 use miden_client::account::AccountId;
 use miden_client::utils::{Deserializable, Serializable, ToHex, hex_to_bytes};
-use miden_faucet_client::requests::MintRequestError;
 use serde::{Serialize, Serializer};
 use sha3::{Digest, Sha3_256};
 
+use crate::pow::PowError;
 use crate::pow::api_key::ApiKey;
 
 /// The size of the encoded challenge in bytes.
@@ -82,10 +82,10 @@ impl Challenge {
 
     /// Decodes the challenge and verifies that the signature part of the challenge is valid
     /// in the context of the specified secret.
-    pub fn decode(value: &str, secret: [u8; 32]) -> Result<Self, MintRequestError> {
+    pub fn decode(value: &str, secret: [u8; 32]) -> Result<Self, PowError> {
         // Parse the hex-encoded challenge string
         let bytes: [u8; CHALLENGE_ENCODED_SIZE] =
-            hex_to_bytes(value).map_err(|_| MintRequestError::MissingPowParameters)?;
+            hex_to_bytes(value).map_err(|_| PowError::InvalidChallenge)?;
 
         // SAFETY: Length of the bytes is enforced above.
         let target = u64::from_le_bytes(bytes[0..8].try_into().unwrap());
@@ -101,7 +101,7 @@ impl Challenge {
         if signature == expected_signature {
             Ok(Self::from_parts(target, timestamp, account_id, api_key, signature))
         } else {
-            Err(MintRequestError::ServerSignaturesDoNotMatch)
+            Err(PowError::ServerSignaturesDoNotMatch)
         }
     }
