+++ crates/client/src/requests.rs
@@ -1,4 +1,4 @@
-use miden_client::account::{AccountId, AccountIdError};
+use miden_client::account::AccountId;
 use miden_client::note::NoteId;
 use miden_client::transaction::TransactionId;
 use serde::Serialize;
@@ -6,7 +6,7 @@ use tokio::sync::{mpsc, oneshot};
 
 use crate::types::{AssetAmount, ExplorerUrl, NoteType};
 
-pub type MintResponseSender = oneshot::Sender<Result<MintResponse, MintRequestError>>;
+pub type MintResponseSender = oneshot::Sender<Result<MintResponse, MintError>>;
 pub type MintRequestSender = mpsc::Sender<(MintRequest, MintResponseSender)>;
 
 /// A request for minting to the Faucet.
@@ -40,25 +40,7 @@ impl Serialize for MintResponse {
 }
 
 #[derive(Debug, thiserror::Error)]
-pub enum MintRequestError {
-    #[error("account ID failed to parse")]
-    AccountId(#[source] AccountIdError),
-    #[error("asset amount {0} is not one of the provided options")]
-    AssetAmount(u64),
-    #[error("API key {0} is invalid")]
-    InvalidApiKey(String),
-    #[error("invalid POW solution")]
-    InvalidPoW,
-    #[error("POW parameters are missing")]
-    MissingPowParameters,
-    #[error("server signatures do not match")]
-    ServerSignaturesDoNotMatch,
-    #[error("server timestamp expired, received: {0}, current time: {1}")]
-    ExpiredServerTimestamp(u64, u64),
-    #[error("challenge already used")]
-    ChallengeAlreadyUsed,
-    #[error("account is rate limited")]
-    RateLimited,
+pub enum MintError {
     #[error("faucet supply exceeded")]
     AvailableSupplyExceeded,
 }
